fix: resolve dj-rest-auth logout bug
It turns out that dj-rest-auth has a bug that doesn’t allow users to log out (ref: DRF Rest Auth Issues).

The issue is that the samesite attribute we set to ‘None’ in settings.py (JWT_AUTH_SAMESITE = 'None') is not passed to the logout view.

This means that we can’t log out, but must wait for the refresh token to expire instead.

To fix this issue, implemented a custom logout view to set both cookies to an empty string and pass additional attributes like secure, httponly, and samesite.